# #153: Faster PDF preview generation