--- a/app/build.gradle
+++ b/app/build.gradle
@@ -135,8 +135,6 @@ dependencies {
 
     implementation "androidx.datastore:datastore-preferences:1.0.0"
 
-    implementation "com.github.barteksc:pdfium-android:1.9.0"
-
     implementation "ch.acra:acra-http:5.9.3"
     implementation "ch.acra:acra-dialog:5.9.3"