--- a/app/build.gradle
+++ b/app/build.gradle
@@ -83,7 +83,7 @@ android {
 }
 
 dependencies {
-    implementation 'space.taran:arklib:0.1.0-SNAPSHOT-9df644252f'
+    implementation 'space.taran:arklib:0.1.0-SNAPSHOT-560bfc5d8d'
     implementation fileTree(dir: "libs", include: ["*.jar"])
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7"