--- a/app/src/main/java/space/taran/arklib/Lib.kt
+++ b/app/src/main/java/space/taran/arklib/Lib.kt
@@ -1,5 +1,0 @@
-package space.taran.arklib
-
-external fun pdfThumbnailGenerate(imgData: ByteArray): ByteArray
-
-external fun initialRustLogger()