--- a/app/src/main/java/space/taran/arknavigator/mvp/model/repo/preview/generator/LinkPreviewGenerator.kt
+++ b/app/src/main/java/space/taran/arknavigator/mvp/model/repo/preview/generator/LinkPreviewGenerator.kt
@@ -1,10 +1,7 @@
 package space.taran.arknavigator.mvp.model.repo.preview.generator
 
 import android.graphics.Bitmap
-import android.graphics.BitmapFactory
-import space.taran.arklib.loadLinkPreview
 import java.nio.file.Path
-import kotlin.io.path.pathString
 
 object LinkPreviewGenerator : PreviewGenerator() {
     override val acceptedExtensions = setOf("link")
@@ -18,11 +15,12 @@ object LinkPreviewGenerator : PreviewGenerator() {
     }
 
     private fun generatePreview(source: Path): Bitmap {
-        val previewBytes =
-            loadLinkPreview(source.pathString) ?: error("No image inside .link file")
-        if (previewBytes.isEmpty()) {
-            error("Image inside .link file was empty")
-        }
-        return BitmapFactory.decodeByteArray(previewBytes, 0, previewBytes.size)
+        throw NotImplementedError()
+        //        val previewBytes =
+        //            loadLinkPreview(source.pathString) ?: error("No image inside .link file")
+        //        if (previewBytes.isEmpty()) {
+        //            error("Image inside .link file was empty")
+        //        }
+        //        return BitmapFactory.decodeByteArray(previewBytes, 0, previewBytes.size)
     }
 }