--- a/app/src/main/java/space/taran/arknavigator/mvp/model/repo/preview/generator/PdfPreviewGenerator.kt
+++ b/app/src/main/java/space/taran/arknavigator/mvp/model/repo/preview/generator/PdfPreviewGenerator.kt
@@ -1,12 +1,8 @@
 package space.taran.arknavigator.mvp.model.repo.preview.generator
 
 import android.graphics.Bitmap
-import android.net.Uri
-import android.os.ParcelFileDescriptor
-import space.taran.arknavigator.ui.App
 import space.taran.arklib.pdfPreviewGenerate
 import space.taran.arklib.PreviewQuality
-import java.io.FileInputStream
 import java.nio.file.Path
 
 object PdfPreviewGenerator : PreviewGenerator() {
@@ -21,18 +17,6 @@ object PdfPreviewGenerator : PreviewGenerator() {
     }
 
     private fun generatePreview(source: Path): Bitmap {
-        val page = 0
-
-        val finalContext = App.instance
-
-        val fd: ParcelFileDescriptor? =
-            finalContext
-                .contentResolver
-                .openFileDescriptor(Uri.fromFile(source.toFile()), "r")
-        val stream = FileInputStream(fd?.fileDescriptor).readBytes()
-
-        val bitmap = pdfPreviewGenerate(stream, PreviewQuality.HIGH)
-
-        return bitmap
+        return pdfPreviewGenerate(source.toString(), PreviewQuality.HIGH)
     }
 }