--- a/app/src/main/java/space/taran/arknavigator/ui/activity/MainActivity.kt
+++ b/app/src/main/java/space/taran/arknavigator/ui/activity/MainActivity.kt
@@ -27,6 +27,7 @@ import space.taran.arknavigator.ui.App
 import space.taran.arknavigator.utils.LogTags.MAIN
 import space.taran.arknavigator.utils.LogTags.PERMISSIONS
 import javax.inject.Inject
+import space.taran.arklib.initialRustLogger
 
 class MainActivity : MvpAppCompatActivity(), MainView {
 
@@ -55,6 +56,9 @@ class MainActivity : MvpAppCompatActivity(), MainView {
     override fun init() {
         Log.d(MAIN, "initializing")
         System.loadLibrary("arklib")
+        initialRustLogger()
+
+        setSupportActionBar(binding.toolbar)
         binding.bottomNavigation.setOnApplyWindowInsetsListener(null)
         binding.bottomNavigation.setOnItemSelectedListener { item ->
             when (item.itemId) {