#153 
Adopting new PDF functionality from [arklib](https://github.com/ARK-Builders/arklib/)
See https://github.com/ARK-Builders/arklib/issues/11 and https://github.com/ARK-Builders/arklib-android/issues/7 for details