@kirillt 2022-07-10
Awesome that we touched base with and wired some advanced stuff from Rust side to Kotlin.
But here is performance degradation, at the moment.

Manual testing on a folder with many big PDF files produced the following results:
* this branch: 115 seconds
* `main` branch: 40 seconds

The folder contains 469 documents, size of each is from 41KB to 21MB.

Detailed rendering logs:
[pdf-previews_this.txt](https://github.com/ARK-Builders/ARK-Navigator/files/9079840/pdf-previews_this.txt)
[pdf-previews_main.txt](https://github.com/ARK-Builders/ARK-Navigator/files/9079841/pdf-previews_main.txt)