@kirillt 2022-07-10
Some of documents are rendered faster with this version, and some are faster with `main` version.
Yes, it also depends on configurable quality, but here are refined logs for side-by-side comparison:
[this.log](https://github.com/ARK-Builders/ARK-Navigator/files/9079901/this.log)
[main.log](https://github.com/ARK-Builders/ARK-Navigator/files/9079902/main.log)