@kirillt 2022-07-10
Here is extensive, readable report of 383 files:
[pdf-performance.txt](https://github.com/ARK-Builders/ARK-Navigator/files/9079952/pdf-performance.txt)

Previews for these 7 files are rendered faster by new version:
```
/storage/emulated/0/Papers/biomath/05-14_mem_module.PDF
`this`: 1287 ms
`main`: 1535 ms
    Nice! `this` is 19.26% faster than `main`
--
/storage/emulated/0/Papers/cognitive/s41235-021-00334-w.pdf
`this`: 13932 ms
`main`: 15856 ms
    Nice! `this` is 13.80% faster than `main`
--
/storage/emulated/0/Papers/Flash Boys 2.0 - Frontrunning, Transaction Reordering, and Consensus Instability
`this`: 14413 ms
`main`: 15056 ms
    Nice! `this` is 4.46% faster than `main`
--
/storage/emulated/0/Papers/Per Mertin Lof - The Definition of Random Sequences.pdf
`this`: 2794 ms
`main`: 6616 ms
    Nice! `this` is 136.79% faster than `main`
--
/storage/emulated/0/Papers/Scalable, transparent, and post-quantum secure computational
`this`: 2668 ms
`main`: 4579 ms
    Nice! `this` is 71.62% faster than `main`
--
/storage/emulated/0/Papers/The Knowledge Complexity of Interactive Proof-System.pdf
`this`: 4342 ms
`main`: 5655 ms
    Nice! `this` is 30.23% faster than `main`
--
/storage/emulated/0/Papers/TrueBit.pdf
`this`: 3237 ms
`main`: 12890 ms
    Nice! `this` is 298.20% faster than `main`
```

Attaching those 7 files here:
[nice.zip](https://github.com/ARK-Builders/ARK-Navigator/files/9079959/nice.zip)

Full archive (487MB):
https://www.taran.space/dev/pdf-test.zip