@nagato-chan 2022-07-11
> Awesome that we touched base with and wired some advanced stuff from Rust side to Kotlin. But here is performance degradation, at the moment.
> 
> Manual testing on a folder with many big PDF files produced the following results:
> 
>     * **this branch: 115 seconds**
> 
>     * **`main` branch: 40 seconds**
> 
> 
> The folder contains 469 documents, size of each is from 41KB to 21MB.

We could switch to Medium Quality for rendering preview. I have tested it and it indicated that it will significantly cut down the time cost while providing previews with similar quality.