@kirillt 2023-01-13
```
java.lang.UnsatisfiedLinkError: No implementation found for long space.taran.arknavigator.mvp.model.repo.index.ResourceIdKt.computeIdNative(long, java.lang.String) (tried Java_space_taran_arknavigator_mvp_model_repo_index_ResourceIdKt_computeIdNative and Java_space_taran_arknavigator_mvp_model_repo_index_ResourceIdKt_computeIdNative__JLjava_lang_String_2)
	at space.taran.arknavigator.mvp.model.repo.index.ResourceIdKt.computeIdNative(Native Method)
	at space.taran.arknavigator.mvp.model.repo.index.ResourceIdKt.computeId(ResourceId.kt:14)
	at space.taran.arknavigator.mvp.model.repo.index.ResourceMeta$Companion.fromPath(ResourceMeta.kt:30)
	at space.taran.arknavigator.mvp.model.repo.index.PlainResourcesIndex$reindexRoot$2.invokeSuspend(PlainResourcesIndex.kt:167)
	at space.taran.arknavigator.mvp.model.repo.index.PlainResourcesIndex$reindexRoot$2.invoke(Unknown Source:8)
	at space.taran.arknavigator.mvp.model.repo.index.PlainResourcesIndex$reindexRoot$2.invoke(Unknown Source:4)
	at kotlinx.coroutines.intrinsics.UndispatchedKt.startUndispatchedOrReturn(Undispatched.kt:89)
	at kotlinx.coroutines.BuildersKt__Builders_commonKt.withContext(Builders.common.kt:169)
	at kotlinx.coroutines.BuildersKt.withContext(Unknown Source:1)
	at space.taran.arknavigator.mvp.model.repo.index.PlainResourcesIndex.reindexRoot$app_release(PlainResourcesIndex.kt:133)
	at space.taran.arknavigator.mvp.model.repo.index.PlainResourcesIndex$reindex$2.invokeSuspend(PlainResourcesIndex.kt:94)
	at space.taran.arknavigator.mvp.model.repo.index.PlainResourcesIndex$reindex$2.invoke(Unknown Source:8)
	at space.taran.arknavigator.mvp.model.repo.index.PlainResourcesIndex$reindex$2.invoke(Unknown Source:4)
	at space.taran.arknavigator.utils.CoroutineUtilsKt$withContextAndLock$2.invokeSuspend(CoroutineUtils.kt:20)
	at kotlin.coroutines.jvm.internal.BaseContinuationImpl.resumeWith(ContinuationImpl.kt:33)
	at kotlinx.coroutines.DispatchedTask.run(DispatchedTask.kt:106)
	at kotlinx.coroutines.internal.LimitedDispatcher.run(LimitedDispatcher.kt:42)
	at kotlinx.coroutines.scheduling.TaskImpl.run(Tasks.kt:95)
	at kotlinx.coroutines.scheduling.CoroutineScheduler.runSafely(CoroutineScheduler.kt:570)
	at kotlinx.coroutines.scheduling.CoroutineScheduler$Worker.executeTask(CoroutineScheduler.kt:750)
	at kotlinx.coroutines.scheduling.CoroutineScheduler$Worker.runWorker(CoroutineScheduler.kt:677)
	at kotlinx.coroutines.scheduling.CoroutineScheduler$Worker.run(CoroutineScheduler.kt:664)
```