@kirillt 2023-01-14
State of the things: after recent changes in `arklib` Navigator from this PR hangs up during rendering previews for PDFs. Sometimes, after 1 document, sometimes after 10.

If we look closer at this thread: https://github.com/ajrcarey/pdfium-render/issues/59, we'll notice that changes in `arklib` are almost identical to the code sample kindly provided by @ajrcarey :) And he put it clearly that this snippet isn't tested and that Pdfium itself is not thread-safe :) Although it could work.

I suggest to do the simplest thing for this moment and ensure that all PDF previews are generated sequentially. We could batch PDF resources into separate set and process it from a dedicated thread/coroutine. @mdrlzy probably you could implement it?