@kirillt 2023-01-14
Also this feature could help us to debug this https://github.com/ARK-Builders/arklib/issues/33