@kirillt 2023-01-16
Sequential previews generation benchmark:
* `main` branch with parallel generation disabled
    build: https://github.com/ARK-Builders/ARK-Navigator/actions/runs/3931502951
* this PR version
    build: https://github.com/ARK-Builders/ARK-Navigator/actions/runs/3925701199

OS: Android 11
Device: Samsung Galaxy Note 10+
Benchmark data: 477 PDF files, 857 MB total

Results (total indexing time, previews/thumbnails generation is the bottleneck):
**main: 1 min 5 sec**
**this PR: 9 min 43 sec**